# Closes #2139: Add benchmark for bigint conversion