--- a/benchmarks/graph_infra/bigint_conversion.perfkeys
+++ b/benchmarks/graph_infra/bigint_conversion.perfkeys
@@ -1,0 +1,4 @@
+bigint_from_uint_arrays Average time =
+bigint_from_uint_arrays Average rate =
+bigint_to_uint_arrays Average time =
+bigint_to_uint_arrays Average rate =