@ronawho 2023-02-13
> if @ronawho or @bmcdonald3 has time next week to get the numbers on 16-node-cs-hdr, that would be awesome!

Looking good, here's 16-node-cs-hdr results

arkouda bigint array to uint arrays:

| size        | before  | Orig PR  | Cur PR |
| ----------- | ------: | -------: | -----: |
| 10**5/node  | 11.24s  |  0.03s   | 0.01s  |
| 10**8/node  | N/A     |  0.39s   | 0.26s  |
| 2**30/node  | N/A     |  5.21s   | 3.82s  |


Full output with current PR:

```
./benchmarks/run_benchmarks.py bigint_conversion -nl 16 --size=$((2**30))
array size = 1,073,741,824
number of trials =  6
>>> arkouda uint arrays from bigint array
numLocales = 16, N = 17,179,869,184
bigint_from_uint_arrays Average time = 2.2505 sec
bigint_from_uint_arrays Average rate = 113.7538 GiB/sec

>>> arkouda bigint array to uint arrays
bigint_to_uint_arrays Average time = 3.8294 sec
bigint_to_uint_arrays Average rate = 66.8514 GiB/sec
```